requirements.txt
accelerate==0.33.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.2.14
dirtyjson==1.0.8
distro==1.9.0
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.23.5
idna==3.8
Jinja2==3.1.4
jiter==0.5.0
joblib==1.4.2
llama-index-core==0.11.3
llama-index-embeddings-openai==0.2.3
llama-index-llms-huggingface==0.3.1
llama-index-llms-huggingface-api==0.2.0
MarkupSafe==2.1.5
marshmallow==3.22.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numpy==1.26.4
openai==1.43.0
packaging==24.1
pillow==10.4.0
psutil==6.0.0
pydantic==2.8.2
pydantic_core==2.20.1
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
safetensors==0.4.4
setuptools==74.0.0
sniffio==1.3.1
SQLAlchemy==2.0.32
sympy==1.13.2
tenacity==8.5.0
text-generation==0.7.0
tiktoken==0.7.0
tokenizers==0.19.1
torch==2.4.0
tqdm==4.66.5
transformers==4.44.2
typing-inspect==0.9.0
typing_extensions==4.12.2
urllib3==2.2.2
wrapt==1.16.0
yarl==1.9.4
python-dotenv
-e .